third_party/WebKit/LayoutTests/bluetooth/server/disconnect/disconnect-twice-in-a-row.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Calling disconnect twice in a row still results in ' +
  '\'connected\' being false.';

bluetooth_test(() =>
  getHealthThermometerDevice()
  .then(() => requestDeviceWithTrustedClick({
    filters: [{services: ['health_thermometer']}]}))
  .then(device => device.gatt.connect())
  .then(gattServer => {
    gattServer.disconnect();
    assert_false(gattServer.connected);
    gattServer.disconnect();
    assert_false(gattServer.connected);
}), test_desc);
</script>
